implement lxc-cinit daemon
This daemon is started by lxc-start command inside the container.
Main purpose is to listen on a socket an execute commands inside
a container, or execute a shell and proxy the pty stream.

It can also open a pty and set the CONSOLE=xyz for init. That way
we can log the output from init (or use a fifo if there is no pty
available).

Further possibility would be to provide additional services like
syslog (useful for application containers).

Note: quilt complains about trailing spaces, so I removed them.

Signed-off-by: Dietmar Maurer <dietmar@proxmox.com>
Signed-off-by: Daniel Lezcano <dlezcano@fr.ibm.com>