create a presetup structure
The different features being developped for lxc shows we need 
to create resources outside of the container and make them
inherited inside the container.

That is the case for the tty and the exec daemon.

In order to not having everything disseminate in the code,
I introduce a presetup structure to be filled with the
different ressources to be inherited. The different fields
have to be passed to the different subsystems.

In the future, if the code keeps going in that way,
the lxc_conf and start.c should be splitted in presetup.c, setup.c 
and conf.c.

Signed-off-by: Daniel Lezcano <dlezcano@fr.ibm.com>